Update expression.md
I was reading through the documentation while editing another file trying to better understand the '?' wildcard function. I came to this page, which seems to describe the '?' wildcard but has a '.' period instead of a question mark as the symbol. The period on the Alternate Language keyboard, is the same key as the question mark so it struck me as suspicious. If this is an erroneous PR, please forgive.