Update restarting.md
Line 153: originally I just removed "the" but re-phrasing as I did is for this reason: Although HDFS, HBase, and ATSD a being starting in the reverse order from shown above, they are listed here in the order that they should actually be started in. The reverse order here is referring to the first order, not the second, thus I separated it into two clauses. Sort of an interesting grammar occurrence there.